Rafactor fields
