+++ src/showcase/components/Form/FormExample.tsx
@@ -2,11 +2,12 @@ import type { ChangeEvent, FormEvent } from 'react'
 import { useRef, useState } from 'react'
 import type { OptionTypeOfSelectField } from '@/layout/components/Form/SelectField'
 import SelectField from '@/layout/components/Form/SelectField'
-import { VerticalItem, VerticalList } from '@/layout/components/VerticalList'
 import { rangeFromOne } from '@/lib/array/range'
 import type { CheckboxValue } from '@/layout/components/Form/CheckboxField'
 import CheckboxGroup from '@/layout/components/Form/CheckboxGroup'
 import RadioGroup from '@/layout/components/Form/RadioGroup'
+import FieldList from '@/layout/components/Form/FieldList'
+import { VerticalItem } from '@/layout/components/VerticalList'
 
 function createOptions(n: number): OptionTypeOfSelectField[] {
   return rangeFromOne(n).map(
@@ -94,7 +95,7 @@ function FormExample() {
     <div>
       <h5>Form example</h5>
       <form onSubmit={handleSubmit}>
-        <VerticalList ref={listRef} style={{ maxWidth: '400px', width: '100%' }}>
+        <FieldList ref={listRef} style={{ maxWidth: '400px', width: '100%' }}>
           <VerticalItem>
             <SelectField
               name="Age"
@@ -106,7 +107,7 @@ function FormExample() {
               optionModalMaxHeight={200}
             />
           </VerticalItem>
-        </VerticalList>
+        </FieldList>
         <CheckboxGroup
           name="activeDays"
           label="Active days"
